requirements.txt
ultralytics==8.3.183
ncnn==1.0.20250503
numpy<2
smbus2==0.5.0
rpi-lgpio==0.6
websockets==15.0.1
Flask==3.1.2
opencv-python==4.8.0.74
picamera2==0.3.30

